refactor(gateway): split dispatch ownership modules
